iPad pop-over support & Usage of edited image
User code can now specify which view controller is used for presenting
the image picker controller. On iPad, pop-overs are now used for
presenting action sheet and image picker controller. The rect used for
that must be specified in the new property popOverPresentRect. If an
edited image is available it is used now
(UIImagePickerControllerEditedImage).